+++ src/server/game/AI/CoreAI/UnitAI.h
@@ -44,8 +44,6 @@ struct AISpellInfoType;
 enum DamageEffectType : uint8;
 enum Difficulty : uint8;
 enum SpellEffIndex : uint8;
-enum class LootItemType : uint8;
-enum class QuestGiverStatus : uint32;
 
 //Selection method used by SelectTarget
 enum SelectAggroTarget
@@ -318,27 +316,8 @@ class TC_GAME_API UnitAI
         static std::unordered_map<std::pair<uint32, Difficulty>, AISpellInfoType> AISpellInfo;
         static void FillAISpellInfo();
 
-        // Called when a player opens a gossip dialog with the creature.
-        virtual bool GossipHello(Player* /*player*/) { return false; }
-
-        // Called when a player selects a gossip item in the creature's gossip menu.
-        virtual bool GossipSelect(Player* /*player*/, uint32 /*menuId*/, uint32 /*gossipListId*/) { return false; }
-
-        // Called when a player selects a gossip with a code in the creature's gossip menu.
-        virtual bool GossipSelectCode(Player* /*player*/, uint32 /*menuId*/, uint32 /*gossipListId*/, char const* /*code*/) { return false; }
-
-        // Called when a player accepts a quest from the creature.
-        virtual void QuestAccept(Player* /*player*/, Quest const* /*quest*/) { }
-
-        // Called when a player completes a quest and is rewarded, opt is the selected item's index or 0
-        virtual void QuestReward(Player* player, Quest const* quest, uint32 opt);
-        virtual void QuestReward(Player* /*player*/, Quest const* /*quest*/, LootItemType /*type*/, uint32 /*opt*/) { }
-
         // Called when a game event starts or ends
         virtual void OnGameEvent(bool /*start*/, uint16 /*eventId*/) { }
-
-        // Called when the dialog status between a player and the creature is requested.
-        virtual QuestGiverStatus GetDialogStatus(Player* player);
 
         virtual void WaypointPathStarted(uint32 /*nodeId*/, uint32 /*pathId*/) { }
         virtual void WaypointStarted(uint32 /*nodeId*/, uint32 /*pathId*/) { }
